requirements.txt
pandas
requests
plotly
pyyaml
